% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.pmsi_has_time}
\alias{.pmsi_has_time}
\title{Check whether a PMSI raw datetime string contains an explicit time component}
\usage{
.pmsi_has_time(x)
}
\arguments{
\item{x}{A vector-like object representing dates/datetimes (character, Date,
etc.). `NULL` values are treated as empty strings.}
}
\value{
A logical vector of the same length as `x`, `TRUE` when the element
  contains a time pattern `\\d{2}:\\d{2}`, otherwise `FALSE`.
}
\description{
Determines if the input contains a time pattern of the form `"HH:MM"`.
This is used to differentiate date-only strings from datetime strings.
}
\examples{
.pmsi_has_time(c("2024-01-01", "2024-01-01 13:45", "01/02/2024 08:00"))
.pmsi_has_time(NA_character_)

}
\keyword{internal}
