% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{edsan_references}
\alias{edsan_references}
\title{Deprecated reference catalog name}
\usage{
edsan_references()
}
\value{
The reference catalog returned by \code{\link[=edsan_reference_catalog]{edsan_reference_catalog()}}.
}
\description{
Deprecated reference catalog name
}
