% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Any object. If not `NULL`, it is returned as-is.}

\item{y}{Any object. Returned only when `x` is `NULL`.}
}
\value{
`x` if `x` is not `NULL`, else `y`.
}
\description{
Returns `x` if it is not `NULL`, otherwise returns `y`.
This is useful for providing defaults when optional values may be missing.
}
\examples{
NULL \%||\% 1
0 \%||\% 1
list(a = 1) \%||\% list(a = 2)

}
\keyword{internal}
