% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doceds.R
\name{process_doceds}
\alias{process_doceds}
\title{Deprecated DOCEDS normalizer name}
\usage{
process_doceds(data)
}
\arguments{
\item{data}{A data frame returned by the DOCEDS data endpoint.}
}
\value{
The value returned by \code{\link[=doceds_normalize]{doceds_normalize()}}.
}
\description{
Deprecated DOCEDS normalizer name
}
