% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doceds-trim-onnx.R
\name{edsan_trimmer_cache_dir}
\alias{edsan_trimmer_cache_dir}
\title{Standard user cache directory for edsan-doc-trimmer}
\usage{
edsan_trimmer_cache_dir()
}
\value{
Path string to the cache folder.
}
\description{
Standard user cache directory for edsan-doc-trimmer
}
